API_Entregas: testes unitários.

+++ src/main/java/br/com/entregas/infraestrutura/adaptador/persistencia/EntregaRepositorioAdaptador.java
@@ -3,6 +3,8 @@ package br.com.entregas.infraestrutura.adaptador.persistencia;
 import java.util.Optional;
 import java.util.UUID;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
 import br.com.entregas.dominio.modelo.EntregaInput;
@@ -15,6 +17,8 @@ import br.com.entregas.infraestrutura.adaptador.persistencia.repositorio.Entrega
 @Repository
 public class EntregaRepositorioAdaptador implements EntregaPersistenciaPorta 
 {    
+	private static final Logger LOGGER = LoggerFactory.getLogger( EntregaRepositorioAdaptador.class );
+	
     private final EntregaJpaRepositorio entregaRepository;
     private final EntregaMapper         entregaMapper;
     
@@ -27,13 +31,17 @@ public class EntregaRepositorioAdaptador implements EntregaPersistenciaPorta
     @Override
     public Optional<EntregaOutput> buscarPorId( UUID id ) 
     {
-        return entregaRepository.findById( id )
+    	LOGGER.debug( "Buscando entrega com ID: {}", id );
+        
+    	return entregaRepository.findById( id )
                 .map( entregaMapper::paraOutput );
     }
 
 	@Override
 	public EntregaOutput salvar( EntregaInput entrega ) 
 	{
+		LOGGER.debug( "Salvando entrega: {}", entrega );
+		
 		EntregaEntidade entidade = entregaMapper.paraEntidade( entrega );
 		
 		entidade = entregaRepository.save( entidade );
@@ -44,12 +52,22 @@ public class EntregaRepositorioAdaptador implements EntregaPersistenciaPorta
 	@Override
     public void deletar( UUID id ) 
 	{
+		LOGGER.debug( "Deletando entrega com ID: {}", id );
+		
+		if( !existePorId( id ) ) 
+		{
+			LOGGER.warn( "Tentativa de deletar entrega inexistente com ID: {}", id );
+			return;
+		}
+		
 		entregaRepository.deleteById(id);
     }
 
     @Override
     public boolean existePorId( UUID id ) 
     {    	
+    	LOGGER.debug( "Verificando existência de entrega com ID: {}", id.toString( ) );
+    	
         return entregaRepository.existsById( id );
     }
 }
